Tidy up production of vax.yaml.
As with refs.info, calling to_yaml() on each collection member produced a set of concatenated yaml statements separated by "---".
On loading that caused the first set of statement to be read but everything else to be ignored. Fix that by making the collection
generate the appropriate yaml when its to_yaml() function is invoked.

scripts/DataSystemsInfo.rb:
  Tidy up the comments.
  Generate "sys-class" in the same way that everything else is generated.
  Provide a to_yaml() function and adjust the encode_with() to produce the expected output.

scripts/systems-info-to-yaml.rb: Invoke to_yaml() on the generated collection rather than on each object in the collection.