Set up so that references make it to the systems YAML file.
This is the first step in getting refs into the mediawiki output.

Makefile: Pass the refs.yaml to the build of a system yaml ...

scripts/systems-info-to-yaml.rb: ... so that the refs.yaml can be loaded and the result passed ...

scripts/DataSystemsInfo.rb: ... to the info file processor. Here the lrefs are processed and turned into
a suitable ID for the appropriate ref.