Move the tags to YAML.
Currently tags are in systems-tags.yaml but this file will probably be extended to handle all tags so the name may change.

Loaded tags end up as an array of hashes. Each hash is a single key (the tag) with a value that is an array.
The array may be nil or it may be of limited length, but if array[1] at least is not present, the tag will be ignored.
array[0] is the text to use when displaying this item.
array[1] is the type of entry for which this tag is valid.
 This might be vax or alpha but it might also be systems (for all systems entries) or all (for all entries).
 This value may be an array in the future (to allow, for example, vax and alpha but nothing else) but currently that is not supported.
array[2] is the sub-heading into which this tag should be grouped (e.g. summary)
array[3] may be a single flag. Only 'required' is currently expected. This means that the tag should be displayed even if it is nil.

DataTags.rb currently only parses the tags and leaves an array of values. In the future it is likely to convert to an array of more descriptive objects.

scripts/systems-tags.yaml: Rather than defining tags in code, do so in YAML.

scripts/DataTags.rb: A noddy parser for the tags YAML.

Makefile: Pass the tags file into the into-to-yaml generator.

scripts/systems-info-to-yaml.rb: Accept a tags YAML file, parse it to find tags and pass these in to the Systems processor.

scripts/DataSystemsInfo.rb: Accept a set of tags. When comparing, check against an uppercase version of the tag (to make the comparison case-blind). Hack out Power/Dimensions/Options until the parsing copes properly. Remove the special case code for "sys-class". Fix some comments.