Makefile: Avoid some tedious repeptition in the makefile while expanding to build all systems files.
scripts/DataTags.rb: Add a Tag object to represent all that matters about a single Tag. Change the loading to build a hash of {names => Tag}. Provide [] to return a named Tag object.

scripts/systems-yaml-to-mediawiki.rb: Change to using the new tag support rather than hard-coding tags in the source.

scripts/systems-tags.yaml: Q-bus is used by MIPS as well as VAX, so allow 'all' to use it. I-inrush is used so allow for it.