Try to verify a sample set of references.
info/vax.info: Have a go at verifying a number of references for the VAX 400-100, VAX 4000-105A and VAX 4000-106A.

scripts/systems-tags.yaml: The same manual specified a "CPU clock" (71MHz) and a "CPU clock rate" (286MHz). It's likely that one is the raw clock supplied to the chip and that the other is the clock used internally by the chip (seemingly 1/4 of the clock rate input) but nothing explicitly makes this clear, so leave both in.