Build all the infobox-based data and try to restrict options to those cases where they can happen.
The restictions should cut down on the amount of text generated in the infobox templates.
For example, Laserbus and XMI were only ever used on VAX or Alpha; VAXBI was only ever used on VAX and MIPS.

Makefile: Build the mediawiki infobox templates and infobox pages for all systems.

scripts/DataTags.rb: Add support for an array of items for applicability.

scripts/systems-tags.yaml: Restrict some options to those cases where they are believed to be possible.