scripts/systems-yaml-to-infobox-mediawiki.rb: Begin to put in code to build XML suitable for import to mediawiki.
Make it possible to write out a file that can be used to import into mediawiki.
To do this wrap everything in <mediawiki> ... </mediawiki> tags.

Each page should be within <page> ... </page> tags. The page text should
be preceded by a <title> ... <title> and be included in <revision> ... <revision> along with a tags describing the comment, the timestamp and the contributor.

The page text should be mediawiki markup but certain characters (the duoble quote, the ampersand and the angle brackets) need to be escaped to avoid interfering with the XML parsing.