Fix potentially ambiguous branch name.
If there happens to be a file called 'master' in the current directory when doing a 'make bundle' the following error will ensue:

[

$ git bundle create bin/info-data.bundle master
fatal: ambiguous argument 'master': both revision and filename
Use '--' to separate paths from revisions, like this:
'git <command> [<revision>...] -- [<file>...]'
error: rev-list died
]

Fix the command to make it clear to git that 'master' is a branch name.
By appending '--' to the command, git knows that what follwos (nothing, in this case) is a filename and everything before is NOT a filename.