Build a publications file.
This is quite similar to a reference file but simply lists publications.
For now no attempt has been made to have any common treatment of pubs or refs, although this may happen in the future.

info/pubs-refs.txt: A set of raw publications, with some duplcaites.

scripts/pub-txt-to-yaml.rb: A basic script to invoke ...

scripts/DataPubTxt.rb: ... the actual text processor ...

Makefile: ... and a few lines in the make file to ensure the yaml pubs file is built.