Add the ability to list related documentation in a systems entry.
scripts/DataSystemsInfo.rb: Parse lines of the form:
 **document = doc{publication-id}
and store them in yaml (as an array called "docs").

scripts/systems-info-to-yaml.rb: This script now needs to know about pubs when parsing an info file.

scripts/systems-yaml-to-mediawiki.rb: This script needs to explicitly ignore "docs".

scripts/systems-yaml-to-infobox-mediawiki.rb: Handle the "docs" yaml value in a special way (other properties have references, this one does not).

Makefile: Sort out PUBS as required.

info/vax.info: Add a bunch of appropriate docs to the VAX-11/780 entry to check that everything works.