Consider keeping lrefs in the order (and with the tag) specified in the info file.
The "**def-lref" lines define a set of local tags to use to identify references for a specific entry. The local tags should be an increasing set of numbers with no gaps, although this is not enforced.
It would be helpful (for manual verification) to keep the tags consistent all the way through to the final output (where possible). So "**def-lref{2}=ref{EK-476AA-TI.A01}" should end up generating a reference [2] in the final wiki markup, for example. Currecntly, however, wiki markup references are numbered in the order that they are encountered, which makes manual checking of the output harder than it needs to be.

scripts/DataSystemsInfo.rb: Store the lrefs hash in the System object. The comment says that it should be an array but currently (and possibly forever) it is a hash. When storing the System object as yaml, store the lrefs hash too.