Fix the build.
The previous commit added "local_references" to the generated yaml. There's no display for such a tag (nor should there be) so it needs to be handled specially.

scripts/systems-yaml-to-infobox-mediawiki.rb: Do not try to output 'local_references'.

scripts/systems-yaml-to-mediawiki.rb: Do not try to output 'local_references'. Throw in some debug in the event of a future tag-display-failure.