Add the bus.modules source.
This used to be called bus.info, but it differs from other .info files so rename it to try to minimise future confusion.

Also rename bus-info-to-mediawiki.rb to bus-module-to-mediawiki.rb.

Note that this script is was initially run manually and used to produce a wiki page. That page has since been manually edited. So currently this script is not used. It is kept here in case it proves useful for other module sets that may crop up in the future.

info/bus.module: The raw module data.

scripts/bus-info-to-mediawiki.rb -> scripts/bus-module-to-mediawiki.rb: A rename of the script that processes bus.module. Some comments were lightly edited to change "info" references to "module".