pending/software.txt: Track DCU software releases here.
It's not clear whether it is worth tracking software releases mentioned in DCUs (or whether it might be better to just record the TOC on each CONOLD, for example). As it is easy to do so, record the entries here to save having to go through the various DCUs again should the data prove to be useful in the future.