info/tapes.info: Clarify names for TKZ08 and TKZ09 and data for the TZK11.
The TKZ08 and TKZ09 are 8mm cartridge drives. The names seem to be as now stated, although one occurrence of an inverted K/Z does occur in a DCU that uses the other (believed correct) variant in all other instances.

The TZK11 is a 5.25" QIC drive. In this case I've checked an orignal manual, where the naming is consistent.

The TZK10 naming is as yet to be clarified as the referenced manual is not to hand, so no changes have been made.