Finish preliminary data gathering from DCU issues up to 1996-05-29 (presumed).
I only have access to electronic copies of DCU issues from 1991-09-06 to 1996-05-29 (this latter presumed, based on the date of the entry on EISNER).

I've gone through all the issues and gathered data. Where data wasn't in a form that could be immediately used but might be useful in the future, it has been added to an appropriate file in pending/.

Software has ended up in pending/software.txt, just in case it ever becomes useful.