Report all possible fatal errors before failing
scripts/DataStorageInfo.rb: Report failures one after another and only raise a fatal exception at the very end. This makes it easier to fix as many errors as possible (in a new info file) without having to run the build multiple times.

Currently this only applies for processing of storage info files, but if successful it could be applied to systems info files processing and also future processing.