Implement parsing of terminal entries from info to YAML
The parsing in DataTerminalsInfo.rb has been re-cast to allow for possible (near-) future parsing of Power and Dimensions blocks. It might be possible to enhance the parsing to allow any type of entry with just some judicious changes. For now, this parses terminals.info in the expected way, so it is a proof of concept for the new parsing style.

scripts/VariableWithReference.rb: Move VariableWithreference to its own source file.

scripts/terminals-info-to-yaml.rb: Straightforward mod of the corresponding storage script to handle turning terminal .info file into YAML.

scripts/DataTerminalsInfo.rb: This is where the bulk of the intelligence lives.