scripts/DataTerminalsInfo.rb: Add support for text blocks
Add support for '**Text-start' and '**Text-end'.

There are no examples of this in the current terminals.info, so this was tested by temporarily adding such a block from vax.info to one of the existing entries.

DataTerminalsInfo.rb is the development ground for the new style of block processing, so it needs to support everything that is currently supported for systems and storage.