info/disks-sdi.info: Eliminate irrelevant 'Location' entries
info/disks-sdi.info: Both the RA71 and RA72 entries specified a 'Location' value which looks like a path. This was probably related to, or at least similar to, the 'Html-target' value. The latter was previously used when this data drove a static site generator. As that's no longer relevant, simply remove these two values.

'Html-target' continues to survive simply because there are many more instances of it in the data.