Write Options blocks out to YAML
scripts/InfoFileHandlerOptions.rb: Write the existing options blocks out to YAML.

The original intention seems to have been to link from options to the data (and hence page) related to that option. That seems to be problematic at the moment so simply write the information a-is for now and work on improving this later.