Ensure that every monitors entry has Sys-name and Decs-name
The options code will need to refernce an item by name (e.g. RF31T), which is going to be the Sys-name, and optionally pick up a description, which is going to be the Desc-name. So it makes sense to ensure that these are always present. Do that for monitors.info to start with.