Prepare data for Options to use
Given a YAML file that contains *all* entries, use it to prepare for use by Options lines in other entries.

scripts/entry-yaml-to-infobox-mediawiki.rb: Process the YAMl of all entries to produce a hash  of "name" => "descriptive text". This is not yet used, but is a preparatory step towards proper processing.

Makefile: Construct a YAML file of all entries. Do this by concatenating all generated YAML files (that contain entries), but remove the first line. This first line is "---" and if it appears more than once in a YAML file, processing stops. I tried adding "..." at the end of each generated YAML file, but that didn't help.

This YAML file of all entries is not fed to the script that builds infobox.mediawiki.txt files.