Makefile: Ensure that failed rules produce no output
Use the special target DELETE_ON_ERROR to ensure that a rule that fails will produce no output.

Without this a ruby build script could raise an error and correctly signal a fatal error, but as the output file is often produced via redirection of stdout, the result will be an empty output file. The next invocation of make will see this more recent output and decide (correctly, but unfortunately) that this target does not need to be rebuilt.

Adding the special target is the simplest way to fix this anomaly.