info/optical-scsi.info: Add a new reference
I thought that I could convert some **lref to **vref , but the existing reference is hard to locate and the new one didn't have the data I expected it to have. Might as well refer to it anyway though.

info/optical-scsi.info: Use a new reference

info/refs.info: Define the new reference