Update some Multia information
There is some Multia VX40 info in the APR 1995 SOC Supplement, but there is also a service guide for the Multia that covers the VX40, VX41 and VX42, so use that and add extra info from the SOC

info/refs.info: Update the Multia service manual reference with a link and a date.

scripts/systems-tags.yaml: Add BUS-PCMCIA, as the Multia needs int.

info/alpha.info: Update the multia entry for the VX40 and add further entries for the VX41 and VX42.