Add sanity checks on the part-no field
When editing some references I noticed one occurrence of a part number that ended in "}". This can easily happen as the part number is most easily created by copying and pasting from the start-ref line and eliminating the final "}"; forgetting to do that could easily happen.

scripts/DataPubTxt.rb:
scripts/DataRefInfo.rb: Ensure that the part-no field in refs and pubs does not end in a "}" character. Currently that is not valid in a part number, and is unlikely to ever be valid.

info/refs.info: Fix up the multiple erroneous final "}" in part numbers exposed by the check added here.