scripts/pubs-summary.rb: Script to produce 'document=' entries
This script was written some time ago to take a PUBS file and produce a set of 'document=' entries for each document described in the PUBS file. It helps sometimes when trying to add associated documents to an INFO entry. Commit it as is so that there is a baseline available for further enhancements.