Sort out a few PUBs with missing titles
An earlier build failed because a 'document=' line referred to a publication with a missing title. Add some code to spot those and fix the first few 9or move them out of the way where there is no obvious title). However, there are still 214 PUBs without titles and that's a bit too much to fix manually in one go. Leave the code in that spots them but make it inactive 9and marked with a TODO).

info/pubs-refs.txt: Move a few PUBs to the bottom of the file where they will now not be processed. Add titles to a few others where that is easy to do.

scripts/DataPubTxt.rb: Add in some code to catch a PUB that has a missing title. As there are too many such PUBs (~214) to fix right now, make the code inactive.