Preparatory work for tag value validation
Some tags have values that could potentially be validated. Take one smal step towards being able to do this. Pass the full set of tags in to the outer handler and verify that nothing changes in the output as a result of this small change.

scripts/DataEntriesInfo.rb: Keep hold of the full tag map and generate the array of permitted tags that the rest of the existing code requires.

scripts/DataTags.rb: 'required' is a bad name, it really means 'always display this value, even if it is blank'. Ajust the name to be clearer.
Add 'validate', which will eventually be true when validation is performed, and  'validator', which will indicate quite how to perform the validation. For now the former will be false and the latter nil.

scripts/entries-info-to-yaml.rb: Pass in the tap map and not simply an array of permitted tags.