Baby steps towards data validation
Passing tags to the handlers is not the only issue. There needs to be a way of going from tag as text to tag object; the obvious way to do that is using a map but currently the tags are just an array. So more work will be required.
For now definition of a validator works and ones that specify an invalid method are rejected (although two of the three methods have not been written yet).

scripts/DataTags.rb: Add a validator to each tag; the validator understands how this tag should be checked (assuming it needs validation at all).
The "Yes/No" validator has been written but has not been checked yet.

scripts/systems-tags.yaml: Compatibility-mode looks like something that should only ever be "Yes" or "No", so start with that one.

scripts/DataEntriesInfo.rb: Tags need to be passed into the handlers.