Add a script to check for offline references
scripts/check-for-offline-references.rb: This script processes a .INFO file and checks the results for non-existent references and for references that have no associated URL.

The former condition should never happen. The latter does happen and this script should help sort out those references that *are* online but for which no URL has been recorded.

scripts/DataEntriesInfo.rb: Quick hag to make tags available but thise will be tweaked to be more manageable soon.

scripts/entries-info-to-yaml.rb: Slight tidy up but no functional change.