Do not drop valid ref lines that have a "!" in a URL
scripts/DataRefInfo.rb: The code was trying to drop lines where the first non-whitespace character was an exclamation mark ("!"). It was actually dropping any line with an exclamation mark in it. That caused the URL for DEC-PR-1992-10-26 ("https://groups.google.com/forum/#!topic/biz.dec/bSMGP6gEYSk"), as well as a handful of similar references, to be dropped. That subsequently caused the check for offline references to flag a few references taken from google groups to flag some refs which are clearly online.