Add "int", "float" and "int-range" support
"int" matches a positive decimal integer.
"float" matches a simple FP value without an exponent (e.g. 1.23).
"int-range" matches "N - M" where N and M are positive decimal integers.

Use some of these in system-tags.yaml.

"int-range" was intended for the Num-proc field but that also has instances of just "1" for some uniprocessor systems and also "2 or 4" for the Firefox series. So that will have to be tackled in a slightly different way later on.