Add DEMNA, DEQRA, DENMA, DEWGB and DSRVG
There is a small possibility of confusion.
The DENMA is an XMI ethernet interface.
The DENMA is the DECagent 90 (a DEChub 90 module).
I've double checked and those do seem to be the same apart from the N/M swap!

info/boards.info: Add the DEMNA (XMI) and DEQRA (Qbus) ...

info/comms.info: ... and the DECagent 90, DECbridge 90 and DECserver 90TL+.

info/refs.info: New references and one URL addition.