Allow a reference to specify alternate URls
Each reference should specify a specific source but for online references sometimes an alternate URL makes sense.
One example that has cropped up recently is DEC press releases which now appear in multiple Usenet archives as well as on systems such as EISNER.
By allowing a number of "alternate" URLs it is possible to specify multiple sources so that if one goes disappears, others may be linked to instead.

For the moment these alternate URLs are stored but do not generate any additional links in generated web pages or wiki pages.

scripts/DataRefInfo.rb: Allow Ref-url-alt to be specified multiple times and record the result in YAML.

info/refs.info: Specify multiple URLs for the specific press release on usenet that triggered this.