Add references from a 20-OCT-1992 DEC press release.
info/refs.info: Add a reference to the press release. Also add a reference for the DECrepeater 90FL Owner's Manual, since that is mentioned and there was previously nothing recorded for it.

info/comms.info: Note the announcement date for the products announced in the press release. Also fill out the new entry for the DECrepeater 90FL Owner's Manual.